Move DisplayGitInfo method to Command class
